+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Vision/RedPipeline.java
@@ -187,8 +187,8 @@ public class RedPipeline extends OpenCvPipeline {
         return centerRect;
     }
 
-    public static double getRectPos() {
-        double rectPos = 0;
+    public static int getRectPos() {
+        int rectPos = 0;
 
         if (largestRect != null) {
             if (getCenter() < IMG_WIDTH / 5) {
